Renamed secure-gateway as gateway
