+++ CI/codefresh.yml
@@ -8,11 +8,11 @@ steps:
     commands:
       - ./CI/versions_to_deploy.sh ${{CF_VOLUME_PATH}}/env_vars_to_export  
 
-  BuildSecureGateway:
-    title: Building Secure Gateway
+  BuildGateway:
+    title: Building Gateway
     type: build
-    image_name: soluto/tweek-secure-gateway
-    working_directory: ./services/secure-gateway/
+    image_name: soluto/tweek-gateway
+    working_directory: ./services/gateway/
           
   BuildingGitRepository:
     title: Building Git Repository Docker Image
@@ -72,7 +72,7 @@ steps:
     type: composition
     composition: ./CI/docker-compose.yml
     composition_variables:      
-      - SECUREGATEWAY_IMAGE=${{BuildSecureGateway}}
+      - GATEWAY_IMAGE=${{BuildGateway}}
       - REPOSITORY_IMAGE=${{BuildingGitRepository}}
       - PUBLISHING_IMAGE=${{BuildingPublishing}}
       - API_IMAGE=${{BuildingApiImage}}
@@ -146,7 +146,7 @@ steps:
     type: composition
     composition: ./CI/docker-compose.yml
     composition_variables:      
-      - SECUREGATEWAY_IMAGE=${{BuildSecureGateway}}
+      - GATEWAY_IMAGE=${{BuildGateway}}
       - REPOSITORY_IMAGE=${{BuildingGitRepository}}
       - PUBLISHING_IMAGE=${{BuildingPublishing}}
       - API_IMAGE=${{BuildingApiImage}}
