+++ services/gateway/.dockerignore
