+++ services/gateway/audit/auditor.go
