+++ services/gateway/audit/logAuditor.go
