+++ services/gateway/proxy/proxy.go
@@ -5,7 +5,7 @@ import (
 	"net/http"
 	"net/url"
 
-	"github.com/Soluto/tweek/services/secure-gateway/security"
+	"github.com/Soluto/tweek/services/gateway/security"
 
 	"github.com/urfave/negroni"
 	"github.com/vulcand/oxy/forward"
