+++ services/gateway/security/ResourceDeffinition.md
