+++ services/gateway/transformation/health.go
@@ -6,9 +6,9 @@ import (
 
 	"github.com/urfave/negroni"
 
-	"github.com/Soluto/tweek/services/secure-gateway/appConfig"
-	"github.com/Soluto/tweek/services/secure-gateway/proxy"
-	"github.com/Soluto/tweek/services/secure-gateway/security"
+	"github.com/Soluto/tweek/services/gateway/appConfig"
+	"github.com/Soluto/tweek/services/gateway/proxy"
+	"github.com/Soluto/tweek/services/gateway/security"
 )
 
 // NewHealthHandler return /health endpoint handler
