+++ services/gateway/transformation/transformation.go
@@ -8,9 +8,9 @@ import (
 	"net/url"
 	"regexp"
 
-	"github.com/Soluto/tweek/services/secure-gateway/appConfig"
-	"github.com/Soluto/tweek/services/secure-gateway/proxy"
-	"github.com/Soluto/tweek/services/secure-gateway/security"
+	"github.com/Soluto/tweek/services/gateway/appConfig"
+	"github.com/Soluto/tweek/services/gateway/proxy"
+	"github.com/Soluto/tweek/services/gateway/security"
 	"github.com/gorilla/mux"
 	"github.com/urfave/negroni"
 )
