+++ services/gateway/transformation/transformation_test.go
@@ -9,8 +9,8 @@ import (
 	"regexp"
 	"testing"
 
-	"github.com/Soluto/tweek/services/secure-gateway/appConfig"
-	"github.com/Soluto/tweek/services/secure-gateway/security"
+	"github.com/Soluto/tweek/services/gateway/appConfig"
+	"github.com/Soluto/tweek/services/gateway/security"
 	"github.com/gorilla/mux"
 	"github.com/urfave/negroni"
 )
